xmatch: Don't double-respond to private message with !
This is an issue for G54bot because it does a catchall command for all
the possible commands that could occur during a game, and tells you
about invalid ones.

So if you send a PM with !command, the general matcher picks it up
first, and then the privmsg-only matcher picks it up as well.

So this will refine the matchers: The matcher with the normal prefix
becomes channel-only, the bot-nick prefix becomes channel-only, and then
there is a privmsg-only matcher where ! is optional

This should preserve existing behavior and not double-fire.
The one thing missing is bot-nick prefix via PM, but who would do
that...?