reset_game: Only act on started games
Unstarted games don't really need resets and can just be reset by
kicking all players if necessary.

This is mostly necessary for the waiting room update.

Strictly speaking, resetting an unstarted game is useful if there are a
bunch of users clogging the waiting list who all need to be kicked, but
I'll add it back in if it comes up?